+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ---
 
+#### 31.05.2018
+
+* [Rust] Closest free site & Site's get location (CodinGame)
+
 #### 30.05.2018
 
 * [Rust] Code Royale - Closest site (CodinGame)